Cleanup DeterministicProcessors
