+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/mempool/MempoolRelayTest.java
@@ -23,7 +23,7 @@ import com.radixdlt.application.TokenUnitConversions;
 import com.radixdlt.chaos.mempoolfiller.MempoolFillerModule;
 import com.radixdlt.chaos.mempoolfiller.MempoolFillerUpdate;
 import com.radixdlt.client.ValidatorAddress;
-import com.radixdlt.environment.deterministic.DeterministicConsensusProcessor;
+import com.radixdlt.environment.deterministic.DeterministicProcessor;
 import com.radixdlt.ledger.LedgerAccumulator;
 import com.radixdlt.ledger.SimpleLedgerAccumulatorAndVerifier;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
@@ -57,7 +57,6 @@ import com.radixdlt.counters.SystemCounters.CounterType;
 import com.radixdlt.counters.SystemCountersImpl;
 import com.radixdlt.crypto.ECKeyPair;
 import com.radixdlt.environment.deterministic.ControlledSenderFactory;
-import com.radixdlt.environment.deterministic.DeterministicEpochsConsensusProcessor;
 import com.radixdlt.environment.deterministic.network.ControlledMessage;
 import com.radixdlt.environment.deterministic.network.DeterministicNetwork;
 import com.radixdlt.environment.deterministic.network.MessageMutator;
@@ -145,7 +144,7 @@ public class MempoolRelayTest {
 			.map(Supplier::get)
 			.collect(ImmutableList.toImmutableList());
 
-		this.nodes.forEach(i -> i.getInstance(DeterministicConsensusProcessor.class).start());
+		this.nodes.forEach(i -> i.getInstance(DeterministicProcessor.class).start());
 	}
 
 	private void injectGenesisAtomToThis(ImmutableList<ECKeyPair> nodeKeys) {
@@ -224,7 +223,7 @@ public class MempoolRelayTest {
 		final var nodeIndex = msg.value().channelId().receiverIndex();
 		final var injector = this.nodes.get(nodeIndex);
 		withThreadCtx(injector, () ->
-			injector.getInstance(DeterministicConsensusProcessor.class)
+			injector.getInstance(DeterministicProcessor.class)
 				.handleMessage(msg.value().origin(), msg.value().message(), msg.value().typeLiteral())
 		);
 		return msg;
