+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/recovery/RecoveryLivenessTest.java
@@ -19,7 +19,7 @@ package com.radixdlt.integration.recovery;
 
 import com.google.inject.Provides;
 import com.radixdlt.client.ValidatorAddress;
-import com.radixdlt.environment.deterministic.DeterministicConsensusProcessor;
+import com.radixdlt.environment.deterministic.DeterministicProcessor;
 import com.radixdlt.ledger.LedgerAccumulator;
 import com.radixdlt.ledger.SimpleLedgerAccumulatorAndVerifier;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
@@ -61,7 +61,6 @@ import com.radixdlt.counters.SystemCountersImpl;
 import com.radixdlt.crypto.ECKeyPair;
 import com.radixdlt.environment.EventDispatcher;
 import com.radixdlt.environment.deterministic.ControlledSenderFactory;
-import com.radixdlt.environment.deterministic.DeterministicEpochsConsensusProcessor;
 import com.radixdlt.environment.deterministic.DeterministicSavedLastEvent;
 import com.radixdlt.environment.deterministic.network.ControlledMessage;
 import com.radixdlt.environment.deterministic.network.DeterministicNetwork;
@@ -161,7 +160,7 @@ public class RecoveryLivenessTest {
 		for (Supplier<Injector> nodeCreator : nodeCreators) {
 			this.nodes.add(nodeCreator.get());
 		}
-		this.nodes.forEach(i -> i.getInstance(DeterministicConsensusProcessor.class).start());
+		this.nodes.forEach(i -> i.getInstance(DeterministicProcessor.class).start());
 	}
 
 	boolean mutate(ControlledMessage message, MessageQueue queue) {
@@ -208,7 +207,7 @@ public class RecoveryLivenessTest {
 		this.network.dropMessages(m -> m.channelId().receiverIndex() == index);
 		Injector injector = nodeCreators.get(index).get();
 		stopDatabase(this.nodes.set(index, injector));
-		withThreadCtx(injector, () -> injector.getInstance(DeterministicConsensusProcessor.class).start());
+		withThreadCtx(injector, () -> injector.getInstance(DeterministicProcessor.class).start());
 	}
 
 	private void initSync() {
@@ -239,7 +238,7 @@ public class RecoveryLivenessTest {
 		String bftNode = " " + injector.getInstance(Key.get(BFTNode.class, Self.class));
 		ThreadContext.put("bftNode", bftNode);
 		try {
-			injector.getInstance(DeterministicConsensusProcessor.class)
+			injector.getInstance(DeterministicProcessor.class)
 				.handleMessage(msg.value().origin(), msg.value().message(), msg.value().typeLiteral());
 		} finally {
 			ThreadContext.remove("bftNode");
